requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
absl-py=2.1.0=pypi_0
anyio=4.8.0=pypi_0
argon2-cffi=23.1.0=pypi_0
argon2-cffi-bindings=21.2.0=pypi_0
arrow=1.3.0=pypi_0
asttokens=3.0.0=pypi_0
async-lru=2.0.4=pypi_0
attrs=25.1.0=pypi_0
babel=2.16.0=pypi_0
beautifulsoup4=4.12.3=pypi_0
bleach=6.2.0=pypi_0
bzip2=1.0.8=h2466b09_7
ca-certificates=2024.12.14=h56e8100_0
certifi=2024.12.14=pypi_0
cffi=1.17.1=pypi_0
charset-normalizer=3.4.1=pypi_0
colorama=0.4.6=pypi_0
comm=0.2.2=pypi_0
contourpy=1.3.1=pypi_0
cycler=0.12.1=pypi_0
debugpy=1.8.12=pypi_0
decorator=5.1.1=pypi_0
defusedxml=0.7.1=pypi_0
dtaidistance=2.3.12=pypi_0
exceptiongroup=1.2.2=pypi_0
executing=2.2.0=pypi_0
fastjsonschema=2.21.1=pypi_0
filelock=3.13.1=pypi_0
filterpy=1.4.5=pypi_0
flatbuffers=25.1.24=pypi_0
fonttools=4.55.6=pypi_0
fqdn=1.5.1=pypi_0
fsspec=2024.6.1=pypi_0
future=1.0.0=pypi_0
futures=3.0.5=pypi_0
h11=0.14.0=pypi_0
httpcore=1.0.7=pypi_0
httpx=0.28.1=pypi_0
idna=3.10=pypi_0
ipykernel=6.29.5=pypi_0
ipython=8.31.0=pypi_0
iso8601=2.1.0=pypi_0
isoduration=20.11.0=pypi_0
jax=0.5.0=pypi_0
jaxlib=0.5.0=pypi_0
jedi=0.19.2=pypi_0
jinja2=3.1.5=pypi_0
joblib=1.4.2=pypi_0
json5=0.10.0=pypi_0
jsonpointer=3.0.0=pypi_0
jsonschema=4.23.0=pypi_0
jsonschema-specifications=2024.10.1=pypi_0
jupyter-client=8.6.3=pypi_0
jupyter-core=5.7.2=pypi_0
jupyter-events=0.11.0=pypi_0
jupyter-lsp=2.2.5=pypi_0
jupyter-server=2.15.0=pypi_0
jupyter-server-terminals=0.5.3=pypi_0
jupyterlab=4.3.4=pypi_0
jupyterlab-pygments=0.3.0=pypi_0
jupyterlab-server=2.27.3=pypi_0
kiwisolver=1.4.8=pypi_0
libffi=3.4.2=h8ffe710_5
liblzma=5.6.3=h2466b09_1
libsqlite=3.48.0=h67fdade_1
libzlib=1.3.1=h2466b09_2
markupsafe=3.0.2=pypi_0
matplotlib=3.10.0=pypi_0
matplotlib-inline=0.1.7=pypi_0
mediapipe=0.10.20=pypi_0
mistune=3.1.0=pypi_0
ml-dtypes=0.5.1=pypi_0
mouseinfo=0.1.3=pypi_0
mpmath=1.3.0=pypi_0
nbclient=0.10.2=pypi_0
nbconvert=7.16.6=pypi_0
nbformat=5.10.4=pypi_0
nest-asyncio=1.6.0=pypi_0
networkx=3.3=pypi_0
notebook=7.3.2=pypi_0
notebook-shim=0.2.4=pypi_0
numpy=1.26.4=pypi_0
opencv-contrib-python=4.11.0.86=pypi_0
opencv-python=4.11.0.86=pypi_0
openssl=3.4.0=ha4e3fda_1
opt-einsum=3.4.0=pypi_0
overrides=7.7.0=pypi_0
packaging=24.2=pypi_0
pandas=2.2.3=pypi_0
pandocfilters=1.5.1=pypi_0
parso=0.8.4=pypi_0
pillow=11.1.0=pypi_0
pip=24.3.1=pyh8b19718_2
platformdirs=4.3.6=pypi_0
prometheus-client=0.21.1=pypi_0
prompt-toolkit=3.0.50=pypi_0
protobuf=4.25.6=pypi_0
psutil=6.1.1=pypi_0
pure-eval=0.2.3=pypi_0
pyautogui=0.9.54=pypi_0
pycparser=2.22=pypi_0
pygetwindow=0.0.9=pypi_0
pygments=2.19.1=pypi_0
pymsgbox=1.0.9=pypi_0
pyparsing=3.2.1=pypi_0
pyperclip=1.9.0=pypi_0
pyrect=0.2.0=pypi_0
pyscreeze=1.0.1=pypi_0
pyserial=3.5=pypi_0
python=3.10.16=h37870fc_1_cpython
python-dateutil=2.9.0.post0=pypi_0
python-json-logger=3.2.1=pypi_0
pytweening=1.2.0=pypi_0
pytz=2024.2=pypi_0
pywin32=308=pypi_0
pywinpty=2.0.14=pypi_0
pyyaml=6.0.2=pypi_0
pyzmq=26.2.0=pypi_0
referencing=0.36.2=pypi_0
requests=2.32.3=pypi_0
rfc3339-validator=0.1.4=pypi_0
rfc3986-validator=0.1.1=pypi_0
rpds-py=0.22.3=pypi_0
scikit-learn=1.6.1=pypi_0
scipy=1.15.1=pypi_0
send2trash=1.8.3=pypi_0
sentencepiece=0.2.0=pypi_0
setuptools=75.8.0=pyhff2d567_0
six=1.17.0=pypi_0
sniffio=1.3.1=pypi_0
sounddevice=0.5.1=pypi_0
soupsieve=2.6=pypi_0
stack-data=0.6.3=pypi_0
sympy=1.13.1=pypi_0
terminado=0.18.1=pypi_0
threadpoolctl=3.5.0=pypi_0
tinycss2=1.4.0=pypi_0
tk=8.6.13=h5226925_1
tomli=2.2.1=pypi_0
torch=2.6.0+cu124=pypi_0
torchaudio=2.6.0+cu124=pypi_0
torchvision=0.21.0+cu124=pypi_0
tornado=6.4.2=pypi_0
traitlets=5.14.3=pypi_0
types-python-dateutil=2.9.0.20241206=pypi_0
typing-extensions=4.12.2=pypi_0
tzdata=2025.1=pypi_0
ucrt=10.0.22621.0=h57928b3_1
ultralytics=8.3.65=pypi_0
ultralytics-thop=2.0.14=pypi_0
uri-template=1.3.0=pypi_0
urllib3=2.3.0=pypi_0
vc=14.3=h5fd82a7_24
vc14_runtime=14.42.34433=h6356254_24
vs2015_runtime=14.42.34433=hfef2bbc_24
wcwidth=0.2.13=pypi_0
webcolors=24.11.1=pypi_0
webencodings=0.5.1=pypi_0
websocket-client=1.8.0=pypi_0
wheel=0.45.1=pyhd8ed1ab_1
